+++ .env-sample
@@ -1,3 +1,4 @@
 PORT=3000
 MONGODB_URL="mongodb://localhost:27017/ssher"
 JWT_SECRET="some-random-string-generate-using-jwt-key-generate.js script"
+ENCRYPTION_KEY=""
