+++ public/html/connections.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SSH Client</title>
-    <link rel="stylesheet" href="../css/styles.css">
+    <link rel="stylesheet" href="../css/connections.css">
 </head>
 <body>
     <!-- Logout button -->
@@ -42,6 +42,6 @@
         </div>
     </div>
 
-    <script src="../js/script.js"></script>
+    <script src="../js/connections.js"></script>
 </body>
 </html>
