+++ public/html/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SSHer</title>
-    <link rel="stylesheet" href="/css/styles.css">
+    <link rel="stylesheet" href="../css/index.css">
 </head>
 <body>
     <div id="app">
@@ -27,6 +27,6 @@
             </form>
         </div>
     </div>
-    <script src="/js/script.js"></script>
+    <script src="../js/index.js"></script>
 </body>
 </html>
