+++ utils/encryptionUtil.js
@@ -1,0 +1,25 @@
+// utils/encryptionUtil.js
+
+const crypto = require('crypto');
+
+// Use environment variables for key and algorithm for better security practices
+const algorithm = 'aes-256-cbc';
+const secretKey = process.env.ENCRYPTION_KEY;
+const iv = crypto.randomBytes(16);
+
+const encrypt = (text) => {
+    const cipher = crypto.createCipheriv(algorithm, Buffer.from(secretKey, 'hex'), iv);
+    let encrypted = cipher.update(text, 'utf8', 'hex');
+    encrypted += cipher.final('hex');
+    return `${iv.toString('hex')}:${encrypted}`;
+};
+
+const decrypt = (text) => {
+    const [ivHex, encryptedText] = text.split(':');
+    const decipher = crypto.createDecipheriv(algorithm, Buffer.from(secretKey, 'hex'), Buffer.from(ivHex, 'hex'));
+    let decrypted = decipher.update(encryptedText, 'hex', 'utf8');
+    decrypted += decipher.final('utf8');
+    return decrypted;
+};
+
+module.exports = { encrypt, decrypt };
